delete confirm

+++ public/script.js
@@ -2,8 +2,11 @@
 const btnDelete = document.querySelector("#delete")
 const articleId = btnDelete.getAttribute("data-deleteArticle")
 btnDelete.addEventListener("click", (e) => {
-    fetch(`/user/myArticles/${articleId}`, { method: "DELETE" })
-        .then((response) => response.json())
-        .then((data) => window.location.href = data.articlesLink)
-        .catch((err) => { console.log(err); })
+    const result = confirm('are you sure to delete this article')
+    if (result) {
+        fetch(`/user/myArticles/${articleId}`, { method: "DELETE" })
+            .then((response) => response.json())
+            .then((data) => window.location.href = data.articlesLink)
+            .catch((err) => { console.log(err); })
+    }
 })
